ENG-01: extract _diagnostics (vip, cos2, contributions, RV coefficients)
Move the post-fit model diagnostics and RV/RV2 matrix-correlation helpers
into _diagnostics. PCA|PLS annotations resolve lazily under TYPE_CHECKING
(pointing at the soon-to-be-extracted _pca/_pls). _pca_pls re-imports the
moved names.
